+++ dmr-api/DTO/GlueIngredientDto.cs
@@ -11,9 +11,9 @@ namespace DMR_API.DTO
         public int ID { get; set; }
         public string Name { get; set; }
         public string Code { get; set; }
-        public int Percentage { get; set; }
+        public double Percentage { get; set; }
         public string Position { get; set; }
-        public int Allow { get; set; }
+        public double Allow { get; set; }
         public Ingredient Ingredient { get; set; }
     }
