+++ dmr-api/DTO/GlueIngredientForMapDto.cs
@@ -12,9 +12,9 @@ namespace DMR_API.DTO
         public int GlueID { get; set; }
         public string IngredientName { get; set; }
         public string Code { get; set; }
-        public int Percentage { get; set; }
+        public double Percentage { get; set; }
         public string Position { get; set; }
-        public int Allow { get; set; }
+        public double Allow { get; set; }
         public int ExpiredTime { get; set; }
         public DateTime  CreatedDate { get; set; }
     }
