+++ dmr-api/Models/GlueIngredient.cs
@@ -16,8 +16,8 @@ namespace DMR_API.Models
         public int ID { get; set; }
         public int GlueID { get; set; } // Ma Keo
         public int IngredientID { get; set; } //Thanh phan
-        public int Allow { get; set; }
-        public int Percentage { get; set; } // Ty le phan tram
+        public double Allow { get; set; }
+        public double Percentage { get; set; } // Ty le phan tram
         public string CreatedDate { get; set; } // Ngay pha
         public string Position { get; set; }
         public Glue Glue { get; set; }
